Kontrollfråga

+++ admin/alchemy_supplier_admin.php
@@ -70,7 +70,7 @@ include 'navigation.php';
                 echo "<td>".$supplier->appoximateValue($current_larp)." $currency</td>\n";
                 echo "<td>". showStatusIcon($supplier->allAmountOfIngredientsApproved($current_larp)) ."</td>\n";
                 echo "<td>" . showStatusIcon($supplier->hasDoneWorkshop()) . "</td>\n";
-                echo "<td>" . "<a href='alchemy_supplier_admin.php?operation=delete&Id=" . $supplier->Id . "'><i class='fa-solid fa-trash'></i></td>\n";
+                echo "<td>" . "<a href='alchemy_supplier_admin.php?operation=delete&Id=" . $supplier->Id . "' onclick='return confirm(\"Är du säker på att du vill ta bort $role->Name som lövjerist?\");'><i  class='fa-solid fa-trash'></i></td>\n";
                 echo "</tr>\n";
             }
             echo "</table>";
